Mode `dedupe_sg_schema` out from Socrata into common ingestion routines (isn't specific to it)
