+++ splitgraph/ingestion/socrata/querying.py
@@ -1,7 +1,7 @@
-from typing import Dict, Any, List, Tuple, Optional
+from typing import Dict, Any, Tuple, Optional
 
-from splitgraph.core.sql import POSTGRES_MAX_IDENTIFIER
 from splitgraph.core.types import TableSchema, TableColumn
+from splitgraph.ingestion.common import dedupe_sg_schema
 
 try:
     from multicorn import ANY
@@ -37,52 +37,6 @@ def _socrata_to_pg_type(socrata_type):
         return _socrata_types[socrata_type]
     else:
         return "text"
-
-
-def dedupe_sg_schema(schema_spec: TableSchema, prefix_len: int = 59) -> TableSchema:
-    """
-    Some Socrata schemas have columns that are longer than 63 characters
-    where the first 63 characters are the same between several columns
-    (e.g. odn.data.socrata.com). This routine renames columns in a schema
-    to make sure this can't happen (by giving duplicates a number suffix).
-    """
-
-    # We truncate the column name to 59 to leave space for the underscore
-    # and 3 digits (max PG identifier is 63 chars)
-    prefix_counts: Dict[str, int] = {}
-    columns_nums: List[Tuple[str, int]] = []
-
-    for column in schema_spec:
-        column_short = column.name[:prefix_len]
-        count = prefix_counts.get(column_short, 0)
-        columns_nums.append((column_short, count))
-        prefix_counts[column_short] = count + 1
-
-    result = []
-    for (_, position), column in zip(columns_nums, schema_spec):
-        column_short = column.name[:prefix_len]
-        count = prefix_counts[column_short]
-        if count > 1:
-            result.append(
-                TableColumn(
-                    column.ordinal,
-                    f"{column_short}_{position:03d}",
-                    column.pg_type,
-                    column.is_pk,
-                    column.comment,
-                )
-            )
-        else:
-            result.append(
-                TableColumn(
-                    column.ordinal,
-                    column.name[:POSTGRES_MAX_IDENTIFIER],
-                    column.pg_type,
-                    column.is_pk,
-                    column.comment,
-                )
-            )
-    return result
 
 
 def socrata_to_sg_schema(metadata: Dict[str, Any]) -> Tuple[TableSchema, Dict[str, str]]:
